Move everything in basic_fields.hpp to fields.hpp (API Change)
fix #320
