+++ CHANGELOG.md
@@ -13,6 +13,7 @@ API Changes:
 
 * Provide websocket::stream accept() overloads
 * Refactor websocket decorators
+* Move everything in basic_fields.hpp to fields.hpp
 
 --------------------------------------------------------------------------------
