+++ include/beast/http/impl/fields.ipp
@@ -5,8 +5,8 @@
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 //
 
-#ifndef BEAST_HTTP_IMPL_BASIC_FIELDS_IPP
-#define BEAST_HTTP_IMPL_BASIC_FIELDS_IPP
+#ifndef BEAST_HTTP_IMPL_FIELDS_IPP
+#define BEAST_HTTP_IMPL_FIELDS_IPP
 
 #include <beast/http/detail/rfc7230.hpp>
 #include <algorithm>
